HPLT - 13-DEC-2025. Add Micsoft.Data.Sqlite and int ID to long ID Commit

+++ AppliedAccounts/Data/ReportActionClass.cs
@@ -5,7 +5,7 @@ namespace AppliedAccounts.Data
     public class ReportActionClass
     {
         public ReportType PrintType { get; set; }
-        public int VoucherID { get; set; }
+        public long VoucherID { get; set; }
         public string Vou_No { get; set; }
     }
