+++ AppliedAccounts/Models/PurchaseListModel.cs
@@ -31,12 +31,12 @@ namespace AppliedAccounts.Models
             Records = LoadData();
         }
 
-        public void Print(int _ID)
+        public void Print(long _ID)
         {
             throw new NotImplementedException();
         }
 
-        public void Edit(int _ID)
+        public void Edit(long _ID)
         {
             NavManager!.NavigateTo($"/Purchase/Purchased/{_ID}");
         }
@@ -52,10 +52,10 @@ namespace AppliedAccounts.Models
             {
                 var _Record = new PurchaseRecord
                 {
-                    ID1 = item.Field<int>("ID1"),
+                    ID1 = item.Field<long>("ID1"),
                     Vou_No = item.Field<string>("Vou_No") ?? "",
                     Vou_Date = item.Field<DateTime>("Vou_Date"),
-                    SupplierID = item.Field<int>("Company"),
+                    SupplierID = item.Field<long>("Company"),
                     SupplierTitle = item.Field<string>("CompanyTitle") ?? "",
                     Inv_No = item.Field<string>("Inv_No") ?? "",
                     Inv_Date = item.Field<DateTime>("Inv_Date"),
@@ -64,18 +64,18 @@ namespace AppliedAccounts.Models
                     Remarks = item.Field<string>("Remarks") ?? "",
                     Comments = item.Field<string>("Comments") ?? "",
                     Status = item.Field<string>("Status") ?? "",
-                    ID2 = item.Field<int>("ID2"),
-                    Sr_No = item.Field<int>("Sr_No"),
-                    TranID = item.Field<int>("TranID"),
-                    Inventory = item.Field<int>("Inventory"),
+                    ID2 = item.Field<long>("ID2"),
+                    Sr_No = (int)item.Field<long>("Sr_No"),
+                    TranID = item.Field<long>("TranID"),
+                    Inventory = item.Field<long>("Inventory"),
                     Batch = item.Field<string>("Batch") ?? "",
                     Qty = item.Field<decimal>("Qty"),
                     Rate = item.Field<decimal>("Rate"),
 
-                    TaxID = item.Field<int>("Tax"),
+                    TaxID = item.Field<long>("Tax"),
                     TaxRate = item.Field<decimal>("Tax_Rate"),
                     Description = item.Field<string>("Description") ?? "",
-                    Project = item.Field<int>("Project"),
+                    Project = item.Field<long>("Project"),
 
                 };
                 _result.Add(_Record);
@@ -89,34 +89,34 @@ namespace AppliedAccounts.Models
 
     public class PurchaseRecord
     {
-        public int ID1 { get; set; }
+        public long ID1 { get; set; }
         public string Vou_No { get; set; }
         public DateTime Vou_Date { get; set; }
         public string Inv_No { get; set; }
         public DateTime Inv_Date { get; set; }
         public DateTime Pay_Date { get; set; }
         public string Ref_No { get; set; }
-        public int SupplierID { get; set; }
+        public long SupplierID { get; set; }
         public string SupplierTitle { get; set; }
         public decimal Amount { get; set; }
         public string Remarks { get; set; }
         public string Comments { get; set; }
         public string Status { get; set; }
-        public int ID2 { get; set; }
+        public long ID2 { get; set; }
         public int Sr_No { get; set; }
-        public int TranID { get; set; }
-        public int Inventory { get; set; }
+        public long TranID { get; set; }
+        public long Inventory { get; set; }
         public string Batch { get; set; }
         public decimal Qty { get; set; }
         public decimal Rate { get; set; }
         public decimal Gross => Qty * Rate;
 
-        public int TaxID { get; set; }
+        public long TaxID { get; set; }
         public decimal TaxRate { get; set; }
         public decimal TaxAmount { get; set; }
         public decimal NetAmount => Gross + TaxAmount;
         public string Description { get; set; }
-        public int Project { get; set; }
+        public long Project { get; set; }
 
         /// <summary>
         /// Other DB
